netapplier: retry verification after failing
When failing on verification nmstate is retrying it up to 5 times with a
sleep interval of 1 seconds. This will fix two issues:

    * During the verification sometimes ethtool is reporting 'unknown' in
      duplex state. This is due a lack of time for some NICs to update the
      information. We have noticed it only with Broadcom BCM57840 NetXtreme II
      NIC but it could affect to other NICs.

    * When switching from DHCP without DHCP server to static IPV6(IPv4
      disabled), the nmstate will fail with verification failure showing IPv6
      is still disabled.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>

In addition:

netapplier: do not modify the desired_state during verification

The desired_state is being modified during verification. Instead of
modifying the desired_state, nmstate should make a copy of it and then
modify the copy.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>

In addition:

nm.nmclient: remove delayed quit from the mainloop

The delayed quit is not needed anymore as nmstate is retrying the
verification step up to five times with an sleep interval of one second.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>

In addition:

nm.checkpoint: fix incorrect interval when adding a timer

According to GLib documentation the interval specified when using
timeout_add() is on miliseconds.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>

In addition:

nm sriov: Properly return when device not exists

When interface does not exists, the `nm.sriov._has_sriov_capability()`
will raise:

```python
    def _has_sriov_capability(ifname):
        dev = device.get_device_by_name(ifname)
>       if nmclient.NM.DeviceCapabilities.SRIOV & dev.props.capabilities:
E       AttributeError: 'NoneType' object has no attribute 'props'
```

Fixed by add additional check on whether device is None.

Also fixed the integration test which incorrectly using eth0 which is
not test NIC, changed to eth1.

Signed-off-by: Gris Ge <fge@redhat.com>
Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>

In addition:

nm.sriov: fix typo in sr-iov device capabilities

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>